Add patch to fix gcc5 build failures
With gcc5 (and I guess hardening enabled) all findings of variables
which potentially are used unitilialized get treated as errors and
cause the build to fail.
In this case both findings are false positives as both 8 bit variables
will get 8 bits shifted in, but it does not hurt to initialize those
with 0 anyway.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>