build: Fix linkage of combined shared library
Currently the build for the combined shared library does not store its own
dependencies. Forcing consumers of the .so to know about internal dependencies
and adding them as -l in their compile.
This patch adds the correct linkage for the currently packaged library.
This only applies to DPDK 2.2 if the linker script patches make it into
2.3 and therefore can be dropped when moving to 2.3.
Reference to LP: #1547517