d/t/test-initscripts test for proper allocation of huge pages
The test (as well as the init script) is aware that some environmetns
might not have all hugepage sizes. But in those they are neither
allocated nor checked - yet the dep8 there at least ensures that it is a
graceful handling e.g. via status of init script that still has to be
"active".